+++ packages/whisky/src/provider/maestro/models/asset.rs
@@ -1,8 +1,7 @@
+use super::*;
 use std::collections::HashMap;
 
 use serde::Deserialize;
-
-use crate::provider::maestro::utils::last_updated::LastUpdated;
 
 #[derive(Deserialize, Debug, Clone)]
 pub struct Standards {
