+++ packages/whisky/src/provider/maestro/models/mod.rs
@@ -1,7 +1,17 @@
 pub mod account;
 pub mod address;
 pub mod asset;
+pub mod last_updated;
 pub mod protocol_parameters;
 pub mod script;
 pub mod transaction;
 pub mod utxo;
+
+pub use account::*;
+pub use address::*;
+pub use asset::*;
+pub use last_updated::*;
+pub use protocol_parameters::*;
+pub use script::*;
+pub use transaction::*;
+pub use utxo::*;
