+++ packages/whisky/src/provider/maestro/models/transaction.rs
@@ -1,8 +1,7 @@
 use std::collections::HashMap;
 
+use super::*;
 use serde::Deserialize;
-
-use crate::provider::maestro::utils::last_updated::LastUpdated;
 
 use super::{script::Script, utxo::Utxo};
