+++ packages/whisky/src/provider/maestro/utils/mod.rs
@@ -1,7 +1,13 @@
-pub mod account_utils;
-pub mod address_utils;
-pub mod block_utils;
-pub mod last_updated;
-pub mod protocol_utils;
-pub mod transaction_utils;
-pub mod utxo_utils;
+mod account_utils;
+mod address_utils;
+mod block_utils;
+mod protocol_utils;
+mod transaction_utils;
+mod utxo_utils;
+
+pub use account_utils::*;
+pub use address_utils::*;
+pub use block_utils::*;
+pub use protocol_utils::*;
+pub use transaction_utils::*;
+pub use utxo_utils::*;
